requirements.txt
-i https://pypi.org/simple
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
fastapi==0.101.1
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
mypy==1.5.1 ; python_version >= '3.8'
mypy-extensions==1.0.0 ; python_version >= '3.5'
pydantic==2.2.0 ; python_version >= '3.7'
pydantic-core==2.6.0 ; python_version >= '3.7'
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.20
sqlalchemy-stubs==0.4
starlette==0.27.0 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==2.0.4 ; python_version >= '3.7'
uvicorn==0.23.2
